final flask command restructure
