+++ backend/.gitignore
@@ -1,3 +1,5 @@
+.ruff_cache
+
 # Byte-compiled / optimized / DLL files
 __pycache__/
 *.py[cod]
