+++ backend/oeps/config.py
@@ -1,6 +1,5 @@
 from pathlib import Path
 
-
 CACHE_DIR = Path(__file__).parent.parent / '.cache'
 DATA_DIR = Path(__file__).parent / 'data'
 RESOURCES_DIR = Path(__file__).parent / 'data' / 'resources'
