+++ backend/setup.py
@@ -6,7 +6,6 @@ setup(
     packages=['oeps'],
     install_requires=[
         'requests==2.31.0',
-        'python-dotenv==1.0.0',
         'google-cloud-bigquery==3.11.3',
         'urllib3<2.0',
         'Flask',
@@ -15,5 +14,6 @@ setup(
         'openpyxl',
         'boto3',
         'duckdb',
+        'ruff',
     ]
 )
